Analyze the `yiiunit\framework\validators` namespace using PHPStan (#20875)
