+++ tests/data/validators/models/ValidatorTestMainModel.php
@@ -10,6 +10,9 @@ namespace yiiunit\data\validators\models;
 
 use yiiunit\data\ar\ActiveRecord;
 
+/**
+ * @property int $id
+ */
 class ValidatorTestMainModel extends ActiveRecord
 {
     public $testMainVal = 1;
