+++ tests/data/validators/models/ValidatorTestRefModel.php
@@ -11,9 +11,9 @@ namespace yiiunit\data\validators\models;
 use yiiunit\data\ar\ActiveRecord;
 
 /**
- * @property int id
- * @property string a_field
- * @property int ref
+ * @property int $id
+ * @property string $a_field
+ * @property int $ref
  */
 class ValidatorTestRefModel extends ActiveRecord
 {
