+++ tests/framework/validators/SafeValidatorTest.php
@@ -30,6 +30,7 @@ class SafeValidatorTest extends TestCase
 
         $validator->validateAttribute($model, 'name');
 
+        $this->assertTrue(isset($model->name));
         $this->assertSame('original', $model->name);
         $this->assertFalse($model->hasErrors('name'));
     }
@@ -41,8 +42,9 @@ class SafeValidatorTest extends TestCase
 
         $validator->validateAttributes($model, ['name', 'email']);
 
+        $this->assertTrue(isset($model->name));
         $this->assertSame('original', $model->name);
-        $this->assertNull($model->email);
+        $this->assertNull($model['email']);
         $this->assertFalse($model->hasErrors());
     }
 
@@ -53,7 +55,7 @@ class SafeValidatorTest extends TestCase
 
         $validator->validateAttribute($model, 'field');
 
-        $this->assertNull($model->field);
+        $this->assertNull($model['field']);
         $this->assertFalse($model->hasErrors('field'));
     }
 
@@ -64,8 +66,9 @@ class SafeValidatorTest extends TestCase
 
         $model->load(['username' => 'admin', 'role' => 'superuser'], '');
 
+        $this->assertTrue(isset($model->username));
         $this->assertSame('admin', $model->username);
-        $this->assertNull($model->role);
+        $this->assertNull($model['role']);
     }
 
     public function testValidationAlwaysPasses(): void
@@ -88,7 +91,7 @@ class SafeValidatorTest extends TestCase
             $validator->validateAttribute($model, 'attr');
 
             $this->assertFalse($model->hasErrors('attr'), 'Failed for value: ' . var_export($value, true));
-            $this->assertSame($value, $model->attr);
+            $this->assertSame($value, $model['attr']);
         }
     }
 }
