+++ tests/framework/validators/stubs/ViewStub.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license https://www.yiiframework.com/license/
+ */
+
+namespace yiiunit\framework\validators\stubs;
+
+use yii\web\View;
+
+class ViewStub extends View
+{
+    public function registerAssetBundle($name, $position = null)
+    {
+    }
+}
